eip3529tests: refactor tests
